Fixes for broken methods
The arguments object gets set for each scope, but all implementations appear to be copy-pasted from MDN's polyfill suggestions which do not use generators and thus do not introduce additional scope. The solution is to use the named function argments, which are undefined when not used. In one case checking against undefined works just as well as the length of the arguments object. In the other case the default value for thisArg is undefined anyway, so we can pass it through without issue. The use of "void 0" has also been dropped, but wasn't really the right way to write the code for Node.js anyway since this is a trick for old browsers that didn't define "undefined" as a symbol.

Bonus: Also added jshint configuration and fixed several jshint issues.